+++ packages/create-yoshi-app/templates/flow-editor/typescript/dev/sites.js
@@ -3,5 +3,4 @@ module.exports = {
   viewerUrl:
     'https://www.editor-flow-dev.com/?appDefinitionId={%flowData.appDefinitionId%}&appName={%flowData.appName%}',
   editorUrl: null,
-  sentryUrl: 'https://sentry.wixpress.com/sentry/{%sentryData.projectName%}',
 };